requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
openai==1.59.7
azure-ai-textanalytics==5.3.0
azure-core==1.32.0
scikit-learn==1.6.1
numpy==2.2.2
httpx==0.28.1
python-dotenv==1.0.1
python-docx==1.1.2
PyPDF2==3.0.1
python-multipart==0.0.19
